Fix discovery & refactor
Fix discovery by using the simplest ResolverSequence: just mdns.rst.DNSServiceResolve
that populates the host and port in the service record.
Refactor so that discoverAndConnect uses startDiscovery/stopDiscovery.
Change startDiscovery so that it returns a Promise that resolves
with the first discovery.
Change the returned discovery to include host, port, httpResponse and
service.