Implement subscribe and subscribeAll
subscribe and subscribeAll return a promise and handle WebSocket
connection. connect only establishes connection with the server, which
in the case of Signal K means only that makes an HTTP request to
/signalk and ensures that the server provides an endpoint for the
version of Signal K that the client is expecting.

All of the WebSocket setup happens in subscribe now. subscribeAll is
just a wrapper around subscribe for the most common use case;